# lab6/src/main/resources/api.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Api
  license:
    name: MIT
servers:
  - url: http://localhost:8080/
paths:
...
  /users/login/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: The id of the user to log in
        schema:
          type: long
    post:
      summary: Log in
      operationId: loginUser
      tags:
        - User operations
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserLoginResponse"
        '401':
          description: Login failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '500':
          description: internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /users/logout:
    post:
      summary: Log out from account
      operationId: logoutUser
      tags:
        - User operations
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserLogoutResponse"
        '401':
          description: Login failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /users:
    post:
      summary: Create a user
      operationId: userCreate
      tags:
        - User operations
      requestBody:
        description: Id for a new User
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreateRequest"
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserCreateResponse"
        '500':
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      summary: List all users with their vote counts
      operationId: listUsers
      tags:
        - User operations
      responses:
        '200':
          description: List of users with vote counts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /users/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: The id of the user to find
        schema:
          type: long
          format: int32
    get:
      summary: Info for a specific user
      operationId: getUserById
      tags:
        - User operations
      responses:
        '200':
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        '404':
          description: not found error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '500':
          description: internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
...
    UserResponse:
      type: object
      required:
        - userId
        - votes
      properties:
        userId:
          type: long
        votes:
          type: array
          items:
            type: long
...
    UserLogoutResponse:
      type: object
      properties:
        message:
          type: string
    UserCreateRequest:
      type: object
      required:
        - userId
      properties:
        userId:
          type: long
    UserCreateResponse:
      type: object
      properties:
        message:
          type: string
    UserLoginResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    UserListResponse:
      type: array
      items:
        $ref: "#/components/schemas/UserResponse"
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string